pvalue.systematic {SCRT}	R Documentation

P-value using the systematic procedure

Description

The P-value corresponding to the observed value of the test statistic is obtained by locating this value in the randomization distribution generated by complete enumeration of all assignment possibilities (the exhaustive randomization distribution).

Usage

pvalue.systematic(design, statistic, save = "no", 
limit, data = read.table(file.choose(new = FALSE)), 
starts = file.choose(new = FALSE), assignments = file.choose(new = FALSE))

Arguments

design
Type of single-case design: "AB", "ABA", "ABAB", "CRD" (completely randomized design), "RBD" (randomized block design), "ATD" (alternating treatments design), "MBD" (multiple-baseline AB design) or "Custom" (user specified design).

statistic
Test statistic. For alternation designs, multiple-baseline designs and AB phase designs, there are 3 built-in possibilities: "A-B", "B-A", and "|A-B|", which stand for the (absolute value of the) difference between condition means. For phase designs with more than 2 phases, 3 more built-in options are available: "PA-PB", "PB-PA", and "|PA-PB|" refer to the (absolute value of the) difference between the means of phase means. Additionally, it is possible to specify a custom test statistic using the variable identifiers "A" and "B" (or in the case of phase deisgns with more than 2 phases, "A1", "B1", "A2", "B2", "A" and "B") and any of the basic R functions. For example, "abs(mean(A) - mean(B))" can be used as a test statistic and it will be the same as using "|A-B|".

save
Save the randomization distribution to a file (save="yes") or just see it as output in the R console (default: save="no").

limit
For phase designs: minimum number of observations per phase. For alternating treatments designs: maximum number of consecutive administrations of the same condition.

data
File in which the data can be found. Default: a window pops up in which the file can be selected.

starts
Only for multiple baseline designs: location of the file where the possible start points can be found. Default: a window pops up in which the file can be selected.

assignments
Only for user specified designs: location of the file where all the possible assignments can be found. Default: a window pops up in which the file can be selected.

Details

When using the default data argument, a window will pop up to ask in what file the data can be found. This text file containing the data should consist of two columns for single-case phase and alternation designs: the first with the condition labels and the second with the obtained scores. For multiple-baseline designs it should consist of these two columns for EACH unit. This way, each row represents one measurement occasion. It is important not to label the rows or columns.

For multiple baseline designs, when using the default starts argument, second a window pops up in which is asked in what file the possible start points can be found. In this startpoint file, each row should contain all possibilities for one unit, separated by a tab. The rows and columns should not be labeled.

For user specified designs, when using the default assignments argument, second a window pops up in which is asked in what file all the possible assignments can be found. In this file, each row should contain the sequence of conditions in one possible assignment, separated by a tab. There should be one row for every possible assignment. The rows and columns should not be labeled.

Missing data should be indicated as NA. When there is missing data, randomization distribution is generated as usual, but instead of randomly reshuffling numerical scores only, the missing data markers (NA) are also included in the reshuffling. For test statistic calculations, missing data are omitted. If test statistic cannot be calculated for a particular randomization due to insufficient data for a treatment condition, the test statistic from this randomization is conservatively considered more extreme than the observed test statistic.

When choosing to save the randomization distribution to a file, next a window will pop up (for multiple baseline designs or user specified designs this is the third pop-up window, for all other designs it is the second window) to ask where to save it. This location can be an existing file, as well as a new file that can be created by giving a file name and the extension .txt. In this latter case a confirmation is required ("The file does not exist yet. Create the file?").

References

Bulte, I., & Onghena, P. (2008). An R package for single-case randomization tests. Behavior Research Methods, 40, 467-478.

Bulte, I., & Onghena, P. (2009). Randomization tests for multiple baseline designs: An extension of the SCRT-R package. Behavior Research Methods, 41, 477-485.

Edgington, E.S., & Onghena, P. (2007). Randomization Tests (4th ed.). Boca Raton, FL: Chapman & Hall/CRC.

http://ppw.kuleuven.be/home/english/research/mesrg

See Also

distribution.systematic to generate the corresponding exhaustive randomization distribution.

observed to calculate the observed test statistic.

distribution.random to generate the nonexhaustive randomization distribution and pvalue.random to obtain the corresponding p-value.

Examples

data(ABAB)
pvalue.systematic(design = "ABAB", statistic = "PA-PB", save = "no", 
limit = 4, data = ABAB)

[Package SCRT version 1.3.0 Index]