quantity {SCRT}	R Documentation

Number of assignment possibilities

Description

The number of assignment possibilities for the specified design is calculated.

Usage

quantity(design, MT, limit, starts = file.choose(new = FALSE), 
assignments = file.choose(new = FALSE))

Arguments

design
Type of single-case design: "AB", "ABA", "ABAB", "CRD" (completely randomized design), "RBD" (randomized block design), "ATD" (alternating treatments design), "MBD" (multiple-baseline AB design) or "Custom" (user specified design).

MT
Measurement times: number of observations.

limit
For phase designs: minimum number of observations per phase. For alternating treatments designs: maximum number of consecutive administrations of the same condition.

starts
Only for multiple baseline designs: location of the file where the possible start points can be found. Default: a window pops up in which the file can be selected.

assignments
Only for user specified designs: location of the file where all the possible assignments can be found. Default: a window pops up in which the file can be selected.

Details

For multiple baseline designs, when using the default starts argument, only the design argument is required. In this default version a window pops up in which is asked in what file the possible start points can be found. In this startpoint file, each row should contain all possibilities for one unit, separated by a tab. The rows and columns should not be labeled.

The number of possible assignments for a multiple baseline design is calculated assuming NO overlap between the possible start points of the different units (staggered administration).

For user specified designs, when using the default assignments argument, a window pops up in which is asked in what file all the possible assignments can be found. In this file, each row should contain the sequence of conditions in one possible assignment, separated by a tab. There should be one row for every possible assignment. The rows and columns should not be labeled.

References

Bulte, I., & Onghena, P. (2008). An R package for single-case randomization tests. Behavior Research Methods, 40, 467-478.

Bulte, I., & Onghena, P. (2009). Randomization tests for multiple baseline designs: An extension of the SCRT-R package. Behavior Research Methods, 41, 477-485.

http://ppw.kuleuven.be/home/english/research/mesrg

See Also

assignments to enumerate all assignment possibilities.

selectdesign to randomly select one of the assignment possibilities.

Examples

quantity(design = "ABAB", MT = 24, limit = 4)

[Package SCRT version 1.3.0 Index]